# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API",
        "description": "Laravel API",
        "contact": {
            "email": "QHs4o@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "config API Server"
        },
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/user/detail": {
            "get": {
                "tags": [
                    "Admin/User"
                ],
                "summary": "Get user detail by email, name, or id",
                "description": "Get user detail by email, name, or id",
                "operationId": "getUserDetail",
                "parameters": [
                    {
                        "name": "user",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "description": "Name, email, or ID",
                            "type": "string",
                            "example": "QHs4o@example.com"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "user": {
                                                            "type": "array",
                                                            "items": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "John Doe"
                                                                    },
                                                                    "email": {
                                                                        "type": "string",
                                                                        "example": "QHs4o@example.com"
                                                                    },
                                                                    "display_name": {
                                                                        "type": "string",
                                                                        "example": "John Doe"
                                                                    },
                                                                    "email_verified_at": {
                                                                        "type": "string",
                                                                        "example": "2020-01-01 00:00:00"
                                                                    },
                                                                    "gender": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "active": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "role": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "example": "2020-01-01 00:00:00"
                                                                    },
                                                                    "updated_at": {
                                                                        "type": "string",
                                                                        "example": "2020-01-01 00:00:00"
                                                                    },
                                                                    "deleted_at": {
                                                                        "type": "string",
                                                                        "example": "2020-01-01 00:00:00"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Invalid data"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Unauthorized"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "error": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "User not found"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}